Added changes (v 0.6.2) - Levels are now loaded into heap memory - Added main menu - Added resolution changing functionality in options menu - Added draw distance selector in options menu - Added debug controls - Code restructure - Bug fixes

+++ Game_Raycaster/Player.cpp
@@ -18,7 +18,7 @@ Player::~Player()
 
 }
 
-void Player::setPlayerNewPos(short x, short y)
+void Player::setPlayerNewPos(float x, float y)
 {
     position.x = x;
     position.y = y;
